requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
sentence-transformers==2.2.2
python-multipart==0.0.6
pydantic==2.4.2
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==0.8.11
prometheus-client==0.19.0
python-json-logger==2.0.7
tenacity==8.2.3
aiofiles==23.2.1 